Moved files from Features to Infrastructure and done little fixes.
