+++ Ficek_QuantiApp/Ficek_QuantiApp.xcodeproj/xcuserdata/martinficek.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,7 @@
 		<key>Ficek_QuantiApp.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>3</integer>
 		</dict>
 		<key>Overture (Playground) 1.xcscheme</key>
 		<dict>
